requirements.txt
matplotlib==3.5.3
numpy==1.25.2
pandas==2.2.0
pytorch_lightning==2.0.8
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.12.0
seaborn==0.13.2
torch==2.0.0
torchmetrics==1.1.1
